backend/requirements.txt
Flask
Flask-Cors
requests
python-dotenv
Flask-SocketIO
python-socketio
python-engineio
websockets
websocket-client
pydub
gevent
greenlet
gunicorn
camel-ai
playwright
